Expose shared bidi direction detection
