+++ richtext-markdown/src/commonMain/kotlin/com/halilibo/richtext/markdown/MarkdownRichText.kt
@@ -7,10 +7,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
-import com.halilibo.richtext.markdown.rtl.LocalCompatibilityTextAlignOverride
-import com.halilibo.richtext.markdown.rtl.fillMaxWidthForRtlCompatibility
-import com.halilibo.richtext.markdown.rtl.firstStrongTextDirection
-import com.halilibo.richtext.markdown.rtl.toCompatibilityTextDirection
 import com.halilibo.richtext.markdown.node.AstBlockQuote
 import com.halilibo.richtext.markdown.node.AstCode
 import com.halilibo.richtext.markdown.node.AstEmphasis
@@ -31,6 +27,10 @@ import com.halilibo.richtext.markdown.node.AstText
 import com.halilibo.richtext.ui.BlockQuote
 import com.halilibo.richtext.ui.FormattedList
 import com.halilibo.richtext.ui.RichTextScope
+import com.halilibo.richtext.ui.rtl.LocalCompatibilityTextAlignOverride
+import com.halilibo.richtext.ui.rtl.fillMaxWidthForRtlCompatibility
+import com.halilibo.richtext.ui.rtl.firstStrongTextDirection
+import com.halilibo.richtext.ui.rtl.toCompatibilityTextDirection
 import com.halilibo.richtext.ui.string.InlineContent
 import com.halilibo.richtext.ui.string.MarkdownAnimationState
 import com.halilibo.richtext.ui.string.RichTextDecorations
